network_training

+++ scripts/sonification_initialize_model.py
@@ -1,8 +1,10 @@
 # %%
 import os
 import random
+from re import A
 
 import h5py
+from matplotlib.pylab import f
 import matplotlib.pyplot as plt
 import numpy as np
 import torch
@@ -11,6 +13,7 @@ from matplotlib import cm
 from numpy import random
 from sklearn.linear_model import Ridge
 from tqdm import tqdm
+from scipy.linalg import block_diag
 
 import filter.utils as utils
 from filter.approximations import DenseGaussianApproximations
@@ -37,17 +40,29 @@ num_neurons = 25
 CYCLE_FAST = {
     "x0": np.array([0.5, 0]),
     "d": 1,
-    "w": 2 * np.pi * 2,
-    "Q": np.array([[1e-3, 0.0], [0.0, 1e-3]]),
+    "w": 2 * np.pi * 1.5,
+    "Q": None,
     "dt": dt,
 }
 CYCLE_SLOW = {
-    "x0": np.array([1, 1]),
+    "x0": np.array([0.5, 0]),
     "d": 1,
-    "w": 2 * np.pi * 1,
-    "Q": np.array([[1e-3, 0.0], [0.0, 1e-3]]),
+    "w": 2 * np.pi * 0.5,
+    "Q": None,
     "dt": dt,
 }
+
+
+loading_matrix_fast_name = "./data/loading_matrix_fast.npz"
+loading_matrix_slow_name = "./data/loading_matrix_slow.npz"
+
+param = np.load(loading_matrix_fast_name, allow_pickle=True)
+C_fast, b_fast = block_diag(*param["C"]), param["b"].flatten()
+param = np.load(loading_matrix_slow_name, allow_pickle=True)
+C_slow, b_slow = block_diag(*param["C"]), param["b"].flatten()
+
+loading = np.hstack([C_fast, C_slow])
+b = np.vstack([b_fast, b_slow]).flatten()
 
 
 def generate_loading_matrix(num_neurons, cycle_info, scale=1, noise=5):
@@ -87,38 +102,43 @@ def initialize_loading_matrix():
     loading_matrix_fast_name = "./data/loading_matrix_fast.npz"
     loading_matrix_slow_name = "./data/loading_matrix_slow.npz"
 
-    # if not os.path.exists(loading_matrix_fast_name) or not os.path.exists(
-    #     loading_matrix_slow_name
-    # ):
-    C_fast, b_fast = generate_loading_matrix(
-        num_neurons, CYCLE_FAST, scale=3, noise=0.3
-    )
+    if not os.path.exists(loading_matrix_fast_name) or not os.path.exists(
+        loading_matrix_slow_name
+    ):
+        C_fast, b_fast = generate_loading_matrix(
+            num_neurons, CYCLE_FAST, scale=3, noise=0.3
+        )
 
-    np.savez(loading_matrix_fast_name, C=C_fast, b=b_fast)
+        np.savez(loading_matrix_fast_name, C=C_fast, b=b_fast)
 
-    C_slow, b_slow = generate_loading_matrix(
-        num_neurons, CYCLE_SLOW, scale=3, noise=0.3
-    )
-    np.savez(loading_matrix_slow_name, C=C_slow, b=b_slow)
+        C_slow, b_slow = generate_loading_matrix(
+            num_neurons, CYCLE_SLOW, scale=3, noise=0.3
+        )
+        np.savez(loading_matrix_slow_name, C=C_slow, b=b_slow)
 
 
-def generate_trajectory():
-    dt = cycle_info["dt"]
+def generate_sample():
     reference_cycle = limit_circle(**CYCLE_FAST)
     perturb_cycle = limit_circle(**CYCLE_FAST)
     coupled_cycle = two_limit_circle(reference_cycle, perturb_cycle)
-    fast_trajectory = coupled_cycle.generate_trajectory(50000)
+    z_fast = coupled_cycle.generate_trajectory(20000)
+    y_fast = np.exp(z_fast @ C_fast + b_fast)
+    # Y_fast is size of (20000, 50) moving average of size 20 with step size 20 along axis 0
+    y_fast = np.mean(y_fast.reshape(50, -1, 20), axis=2)
 
     reference_cycle = limit_circle(**CYCLE_SLOW)
     perturb_cycle = limit_circle(**CYCLE_SLOW)
     coupled_cycle = two_limit_circle(reference_cycle, perturb_cycle)
-    slow_trajectory = coupled_cycle.generate_trajectory(50000)
+    z_slow = coupled_cycle.generate_trajectory(20000)
+    y_slow = np.exp(z_fast @ C_slow + b_slow)
+    y_slow = np.mean(y_slow.reshape(50, -1, 20), axis=2)
 
-    return np.hstack([fast_trajectory, slow_trajectory])
+    return torch.tensor(np.vstack([y_fast, y_slow]).T), torch.tensor(
+        np.hstack([z_fast[::20], z_slow[::20]])
+    )
 
 
 def train_network():
-
     bin_sz = 20e-3
     device = "cpu"
     data_device = "cpu"
@@ -128,8 +148,8 @@ def train_network():
     torch.set_default_dtype(default_dtype)
 
     """hyperparameters"""
-    n_inputs = 1
-    n_latents = 2
+    n_inputs = 2
+    n_latents = 8
     n_hidden_current_obs = 128
     n_samples = 25
     rank_y = 2
@@ -139,40 +159,28 @@ def train_network():
     blues = cm.get_cmap("Blues", n_samples)
 
     """data params"""
-    n_trials = 1000
+    n_trials = 10
     n_neurons = 100
-    n_time_bins = 25
-    theta = np.pi / 4
+    n_time_bins = 1000
 
-    A = torch.nn.Linear(n_latents, n_latents, bias=False, device=device).requires_grad_(
-        False
-    )
-    A.weight.data = 0  # TODO update this line
     B = torch.nn.Linear(n_inputs, n_latents, bias=False, device=device).requires_grad_(
         False
     )
-    B.weight.data = 0  # TODO update this line
-    C = torch.nn.Linear(n_latents, n_neurons, device=device).requires_grad_(False)
-    C.weight.data = 0  # TODO update this line
+    B.weight.data = torch.tensor(
+        [[0, -1], [1, 0], [0, -1], [1, 0], [0, -1], [1, 0], [0, -1], [1, 0]]
+    ).type(default_dtype)
     Q_0_diag = torch.ones(n_latents, device=device).requires_grad_(False)
     Q_diag = torch.ones(n_latents, device=device).requires_grad_(False)
     R_diag = torch.ones(n_neurons, device=device).requires_grad_(False)
     m_0 = torch.zeros(n_latents, device=device).requires_grad_(False)
 
     """generate input and latent/observations"""
-    u = torch.rand((n_trials, n_time_bins, n_inputs), device=device)
-    y_gt, z_gt = utils.sample_lds(
-        u,
-        A.weight.data,
-        B.weight.data,
-        C.weight.data,
-        C.bias.data,
-        Q_diag,
-        R_diag,
-        m_0,
-        Q_0_diag,
-        device=device,
-    )
+    u = torch.rand((n_trials, n_time_bins, n_inputs), device=device) * bin_sz
+    y_gt = torch.zeros((n_trials, n_time_bins, n_neurons), device=device)
+    z_gt = torch.zeros((n_trials, n_time_bins, n_latents), device=device)
+    for i in range(n_trials):
+        print(f"trial: {i}")
+        y_gt[i], z_gt[i] = generate_sample()
 
     y_train_dataset = torch.utils.data.TensorDataset(
         y_gt,
@@ -187,14 +195,24 @@ def train_network():
     approximation_pdf = DenseGaussianApproximations(n_latents, device)
 
     """likelihood pdf"""
-    C = LinearPolarToCartesian(n_latents, n_neurons, 4, device=device)
-    C.linear.weight.data = None  # TODO update this line
+    C = LinearPolarToCartesian(
+        n_latents, n_neurons, 4, loading=loading, bias=b, device=device
+    )
     likelihood_pdf = PoissonLikelihood(C, n_neurons, bin_sz, device=device)
     likelihood_pdf = PoissonLikelihood(C, n_neurons, delta=bin_sz, device=device)
 
     """dynamics module"""
+
     # dynamics_fn = utils.build_gru_dynamics_function(n_latents, n_hidden_dynamics, device=device)
-    dynamics_fn = A  # TODO Update this line
+    def A(x):
+        Ax = torch.zeros_like(x)
+        Ax[:, :, 0] = x[:, :, 0] + x[:, :, 0] * (1 - x[:, :, 0] ** 2) * bin_sz
+        Ax[:, :, 1] = x[:, :, 1] + 2 * np.pi * 1.5 * bin_sz
+        Ax[:, :, 2] = x[:, :, 2] + x[:, :, 2] * (1 - x[:, :, 2] ** 2) * bin_sz
+        Ax[:, :, 3] = x[:, :, 3] + 2 * np.pi * 0.5 * bin_sz
+        return Ax
+
+    dynamics_fn = A
     dynamics_mod = DenseGaussianNonlinearDynamics(
         dynamics_fn, n_latents, approximation_pdf, Q_diag, device=device
     )
@@ -274,5 +292,5 @@ if __name__ == "__main__":
     np.random.seed(123)
     torch.manual_seed(123)
 
-    initialize_loading_matrix()
-    # train_network()
+    # initialize_loading_matrix()
+    train_network()
